Module Varray__Utils.Array_backend

type 'a t = 'a array

The type of an array.

type 'a elt = 'a

The type of the elements contained in the array.

val length : 'a t -> int

The number of elements stored in the array.

val get : 'a t -> int -> 'a elt

get t i returns the ith element of the array. The elements are numbered from 0 to length t - 1.

val set : 'a t -> int -> 'a elt -> unit

set t i v modifies t in place, replacing the element at position i with the value v.

val empty : 'a t

An empty array with length 0, similar to the polymorphic constant [| |] .

val create : int -> 'a t

create n returns an array of length n. It is guaranteed that the elements will not be accessed through get.

val erase_at : 'a t -> int -> unit

erase_at t i resets the element at position i. It is an opportunity to free the memory of this value. Since the element is now undefined, this index will not be accessed again until a write is done.

val blit : 'a t -> int -> 'a t -> int -> int -> unit

blit a i b j n copies the elements from the range i,i+n-1 of the array a to the range j,j+n-1 of the array b.